move bench to new file
